# Speedup division by vanishing polynomials for dense polynomials