## Description

In Marlin and many other sumcheck protocols, we divide a polynomial by a vanishing polynomial. Sometimes, we only want the quotient. Sometimes, we need both the quotient and the remainder as part of the protocol. 

The current implementation is as follows:
- Call the domain API for a sparse representation of the vanishing polynomial
- Use the general-purpose `divide_with_q_and_r` to compute the quotient and the remainder.

However, we know that vanishing polynomial has a very special shape. The implementation here leverages this pattern.

Benchmark shows that the division by vanishing polynomials is now faster.

"bls12_381" - evaluate_div_by_vanishing_poly/32768                                                                             
                        time:   [37.766 us 38.054 us 38.374 us]
                        change: [-9.2485% -6.5941% -4.1249%] (p = 0.00 < 0.05)
                        Performance has improved.

"bls12_381" - evaluate_div_by_vanishing_poly/65536                                                                            
                        time:   [343.89 us 351.85 us 360.05 us]
                        change: [-93.768% -93.635% -93.488%] (p = 0.00 < 0.05)
                        Performance has improved.

"bls12_381" - evaluate_div_by_vanishing_poly/131072                                                                             
                        time:   [1.4784 ms 1.5313 ms 1.5837 ms]
                        change: [-91.204% -90.275% -88.971%] (p = 0.00 < 0.05)
                        Performance has improved.

---

Before we can merge this PR, please make sure that all the following items have been
checked off. If any of the checklist items are not applicable, please leave them but
write a little note why.

- [x] Targeted PR against correct branch (master)
- [x] Linked to GitHub issue with discussion and accepted design OR have an explanation in the PR that describes this work.
- [x] Wrote unit tests
- [x] Updated relevant documentation in the code
- [x] Added a relevant changelog entry to the `Pending` section in `CHANGELOG.md`
- [x] Re-reviewed `Files changed` in the GitHub PR explorer